refactor(agent): Phase G — ship run-memory as importable package (pilot)
- New npa/src/npa/agent_backend/ package: modules are shipped to the VM as
  importable files instead of string-substituted into the agent.py f-string.
- Migrate agent_memory: logic moves to agent_backend/memory.py; cli/agent_memory.py
  becomes a re-export shim (import paths/tests unchanged). Bootstrap ships the
  package via heredoc; backend.py imports 'from agent_backend.memory import ...'
  (sys.path) instead of inlining RunMemory.
- Establishes the reversible mechanism for migrating the remaining B-D modules;
  embed stays active for everything not yet migrated. Byte-for-byte behavior.
- Tests: shipped-module compile check + assertion that backend.py no longer inlines
  RunMemory and carries no leftover embed placeholder.
