+++ npa/src/npa/cli/agent.py
@@ -231,22 +231,23 @@ def _embedded_agent_semantic_router_source() -> str:
     return raw
 
 
-def _embedded_agent_memory_source() -> str:
-    """Return agent_memory.py source embedded into the remote agent backend."""
-    import re
+def _shipped_agent_backend_module_source(name: str) -> str:
+    """Return the FULL source of a shipped agent_backend module.
 
-    path = Path(__file__).with_name("agent_memory.py")
-    raw = path.read_text(encoding="utf-8")
-    raw = re.sub(r'^""".*?"""\s*\n', "", raw, count=1, flags=re.DOTALL)
-    raw = re.sub(r"^from __future__ import annotations\s*\n", "", raw)
-    return raw
+    Unlike the embed readers, shipped modules are uploaded to the VM as their own
+    importable files (Phase G), so the source is returned verbatim — docstring and
+    ``from __future__`` line intact — not inlined into the backend f-string.
+    """
+    path = Path(__file__).resolve().parents[1] / "agent_backend" / f"{name}.py"
+    return path.read_text(encoding="utf-8")
 
 
 _AGENT_CHAT_EMBED = "__NPA_AGENT_CHAT_EMBED__"
 _AGENT_ACTIONS_EMBED = "__NPA_AGENT_ACTIONS_EMBED__"
 _AGENT_SIM2REAL_LOOP_EMBED = "__NPA_AGENT_SIM2REAL_LOOP_EMBED__"
 _AGENT_SEMANTIC_ROUTER_EMBED = "__NPA_AGENT_SEMANTIC_ROUTER_EMBED__"
-_AGENT_MEMORY_EMBED = "__NPA_AGENT_MEMORY_EMBED__"
+# Phase G: shipped (uploaded + imported) rather than embedded.
+_AGENT_MEMORY_SHIP = "__NPA_AGENT_MEMORY_SHIP__"
 _AGENT_WORKFLOW_EMBED = "__NPA_AGENT_WORKFLOW_EMBED__"
 _AGENT_ARTIFACTS_EMBED = "__NPA_AGENT_ARTIFACTS_EMBED__"
 _AGENT_ROUTING_EMBED = "__NPA_AGENT_ROUTING_EMBED__"
@@ -1793,7 +1794,7 @@ def _bootstrap_agent_stack(
     agent_actions_source = _embedded_agent_actions_source()
     agent_sim2real_loop_source = _embedded_agent_sim2real_loop_source()
     agent_semantic_router_source = _embedded_agent_semantic_router_source()
-    agent_memory_source = _embedded_agent_memory_source()
+    agent_memory_ship_source = _shipped_agent_backend_module_source("memory")
     agent_workflow_source = _embedded_agent_workflow_source()
     agent_artifacts_source = _embedded_agent_artifacts_source()
     agent_routing_source = _embedded_agent_routing_source()
@@ -1871,6 +1872,11 @@ cat <<'ENV' | sudo tee /opt/npa-agent/public.env >/dev/null
 NPA_AGENT_PUBLIC_URL=https://{host}
 NPA_AGENT_PUBLIC_HOST={host}
 ENV
+sudo mkdir -p /opt/npa-agent/agent_backend
+printf '' | sudo tee /opt/npa-agent/agent_backend/__init__.py >/dev/null
+cat <<'PY' | sudo tee /opt/npa-agent/agent_backend/memory.py >/dev/null
+{_AGENT_MEMORY_SHIP}
+PY
 cat <<'PY' | sudo tee /opt/npa-agent/backend.py >/dev/null
 import json
 import os
@@ -3539,7 +3545,12 @@ def _chat_with_resilience(
 
 {_AGENT_SEMANTIC_ROUTER_EMBED}
 
-{_AGENT_MEMORY_EMBED}
+# Phase G: run memory is a SHIPPED module (uploaded to /opt/npa-agent/agent_backend
+# and imported here) rather than string-substituted into this f-string.
+import sys as _npa_sys
+if "/opt/npa-agent" not in _npa_sys.path:
+    _npa_sys.path.insert(0, "/opt/npa-agent")
+from agent_backend.memory import RunMemory, JsonFileStore, InMemoryStore
 
 {_AGENT_WORKFLOW_EMBED}
 
@@ -6713,7 +6724,7 @@ sudo systemctl restart npa-agent-backend
         .replace(_AGENT_ACTIONS_EMBED, agent_actions_source)
         .replace(_AGENT_SIM2REAL_LOOP_EMBED, agent_sim2real_loop_source)
         .replace(_AGENT_SEMANTIC_ROUTER_EMBED, agent_semantic_router_source)
-        .replace(_AGENT_MEMORY_EMBED, agent_memory_source)
+        .replace(_AGENT_MEMORY_SHIP, agent_memory_ship_source)
         .replace(_AGENT_WORKFLOW_EMBED, agent_workflow_source)
         .replace(_AGENT_ARTIFACTS_EMBED, agent_artifacts_source)
         .replace(_AGENT_ROUTING_EMBED, agent_routing_source)
